requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Transformers and model loading
transformers>=4.30.0

# Optional but recommended for better performance
python-dotenv>=1.0.0
websocket-client>=1.6.0
websockets>=12.0
pymysql>=1.0.3
